refactor(config-manager): 配置管理器文件重命名简化路径引用
改动说明

- 为什么改：
  - 配置管理模块文件命名不够语义化，configManager.ts 名称重复了目录名 config/，导致路径引用过长（如 @/lib/config/configManager.js）
  - 作为配置管理器模块重构的第一步，为后续进一步优化做准备
- 改了什么：
  a. 文件重命名：
      - apps/backend/lib/config/configManager.ts → apps/backend/lib/config/manager.ts
    - apps/backend/lib/config/configManager.test.ts → apps/backend/lib/config/manager.test.ts
    - apps/backend/lib/config/configManager.json5.test.ts → apps/backend/lib/config/manager.json5.test.ts
    - apps/backend/lib/config/configManager.jsonc.test.ts → apps/backend/lib/config/manager.jsonc.test.ts
  b. 全局导入更新（共 51 个文件）：
      - 所有 @/lib/config/configManager.js 导入更新为 @/lib/config/manager.js
    - 涉及模块：WebServer、handlers、services、cli、adapters、middlewares 等
- 影响范围：
  - 核心影响：配置管理模块的导入路径从 @/lib/config/configManager.js 简化为 @/lib/config/manager.js
  - 兼容性：对外 API 保持不变，仅为路径别名层面的重构，不影响运行时行为
  - 测试影响：所有测试文件导入已同步更新，测试用例逻辑保持不变
- 验证方式：
  - 执行 pnpm type:check 确保类型检查通过
  - 执行 pnpm test 确保所有测试用例通过
  - 确认配置管理器功能（加载、更新、保存配置）正常工作
