+++ apps/backend/WebServer.ts
@@ -1,7 +1,7 @@
 import { createServer } from "node:http";
 import type { IncomingMessage, Server, ServerResponse } from "node:http";
-import { configManager } from "@/lib/config/configManager.js";
-import type { MCPServerConfig } from "@/lib/config/configManager.js";
+import { configManager } from "@/lib/config/manager.js";
+import type { MCPServerConfig } from "@/lib/config/manager.js";
 import type { EndpointConnection } from "@/lib/endpoint/connection.js";
 import { EndpointManager } from "@/lib/endpoint/index.js";
 import type {
