+++ apps/backend/__tests__/config-sync-integration.test.ts
@@ -5,8 +5,8 @@
 
 import { existsSync, readFileSync, writeFileSync } from "node:fs";
 import { resolve } from "node:path";
-import type { AppConfig } from "@/lib/config/configManager.js";
-import { ConfigManager } from "@/lib/config/configManager.js";
+import type { AppConfig } from "@/lib/config/manager.js";
+import { ConfigManager } from "@/lib/config/manager.js";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 // Mock fs module
