+++ apps/backend/adapters/__tests__/integration.test.ts
@@ -6,7 +6,7 @@
 import type {
   LocalMCPServerConfig,
   SSEMCPServerConfig,
-} from "@/lib/config/configManager.js";
+} from "@/lib/config/manager.js";
 import { MCPTransportType } from "@/lib/mcp/types";
 import { describe, expect, it } from "vitest";
 import {
