+++ apps/backend/handlers/ToolApiHandler.ts
@@ -3,11 +3,11 @@
  * 处理通过 HTTP API 调用 MCP 工具的请求
  */
 
-import { configManager } from "@/lib/config/configManager.js";
+import { configManager } from "@/lib/config/manager.js";
 import type {
   CustomMCPTool,
   ProxyHandlerConfig,
-} from "@/lib/config/configManager.js";
+} from "@/lib/config/manager.js";
 import { MCPCacheManager } from "@/lib/mcp";
 import type { MCPServiceManager } from "@/lib/mcp";
 import type { JSONSchema } from "@/lib/mcp/types.js";
