+++ apps/backend/lib/config/manager.json5.test.ts
@@ -4,8 +4,8 @@ import { resolve } from "node:path";
 import JSON5 from "json5";
 import * as json5Writer from "json5-writer";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { AppConfig, ConnectionConfig } from "./configManager";
-import { ConfigManager } from "./configManager";
+import type { AppConfig, ConnectionConfig } from "./manager";
+import { ConfigManager } from "./manager";
 
 // Mock fs module
 vi.mock("node:fs", () => ({
