+++ apps/backend/lib/config/manager.test.ts
@@ -12,8 +12,8 @@ import type {
   CustomMCPTool,
   MCPServerConfig,
   MCPToolConfig,
-} from "./configManager";
-import { ConfigManager } from "./configManager";
+} from "./manager";
+import { ConfigManager } from "./manager";
 
 // Mock fs module
 vi.mock("node:fs", () => ({
