+++ apps/backend/lib/mcp/__tests__/custom.timeout.test.ts
@@ -3,7 +3,7 @@
  * 测试基本功能和资源管理
  */
 
-import type { CustomMCPTool } from "@/lib/config/configManager.js";
+import type { CustomMCPTool } from "@/lib/config/manager.js";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { CustomMCPHandler } from "../custom.js";
 
@@ -28,7 +28,7 @@ vi.mock("@root/Logger.js", () => ({
 }));
 
 // Mock configManager
-vi.mock("../../configManager.js", () => ({
+vi.mock("@/lib/config/manager.js", () => ({
   configManager: {
     getCustomMCPTools: vi.fn(),
     getCustomMCPConfig: vi.fn(),
