+++ apps/backend/lib/mcp/manager.ts
@@ -7,8 +7,8 @@
  */
 
 import { EventEmitter } from "node:events";
-import type { MCPToolConfig } from "@/lib/config/configManager.js";
-import { configManager } from "@/lib/config/configManager.js";
+import type { MCPToolConfig } from "@/lib/config/manager.js";
+import { configManager } from "@/lib/config/manager.js";
 import { MCPService } from "@/lib/mcp";
 import { MCPCacheManager } from "@/lib/mcp";
 import type {
