+++ apps/backend/lib/mcp/transports/__tests__/WebSocketAdapter.test.ts
@@ -10,7 +10,7 @@ import { ConnectionState } from "../TransportAdapter.js";
 import type { WebSocketConfig } from "../WebSocketAdapter.js";
 import { WebSocketAdapter, WebSocketState } from "../WebSocketAdapter.js";
 // Mock ConfigManager 模块，使用项目标准路径别名
-vi.mock("@/lib/config/configManager.js", () => ({
+vi.mock("@/lib/config/manager.js", () => ({
   configManager: {
     configExists: vi.fn().mockReturnValue(true),
     getConfig: vi.fn().mockReturnValue({
